+++ ChangeLog.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
-## [0.9.1] - 2018-08-xx
+## [0.9.1] - 2018-08-30
 
 ### Added